refactor: rework spimi index construction

+++ shared/__init__.py
@@ -1,5 +1,5 @@
 import math
 
 
-def weight(n: int, tf: int, df: int) -> float:
+def weight(n: int, tf: float, df: int) -> float:
     return math.log(1 + tf) * math.log((n + 1) / (df + 1))
